backend/requirements/development.txt
-r base.txt

# Development tools
django-debug-toolbar==4.4.6
django-extensions==3.2.3
ipython==8.30.0

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1
faker==33.1.0

# Code quality
black==24.10.0
flake8==7.1.1
isort==5.13.2
mypy==1.13.0
django-stubs==5.1.1
djangorestframework-stubs==3.15.1
